Culler pseudo compiling
