+++ Sources/AGEngine/Utils/Containers/LFQueue.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <concurrentqueue/concurrentqueue.h>
+
+namespace AGE
+{
+	template <typename T>
+	class LFQueue : public moodycamel::ConcurrentQueue<T>
+	{
+
+	};
+}
